+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.0.1
 
-- The Learmond Corporation's official flutter package. 
+- The Learmond Corporation's CLI tool. 
 - Initial version.